MISRA 10.4

+++ board/config.h
@@ -44,7 +44,7 @@
 #define GET_BUS(msg) ((msg)->bus)
 #define GET_LEN(msg) ((msg)->len)
 #define GET_ADDR(msg) ((msg)->addr)
-#define GET_BYTE(msg, b) ((msg)->data[(int)(b)])
+#define GET_BYTE(msg, b) ((int)(msg)->data[(int)(b)])
 #define GET_BYTES_04(msg) ((msg)->data[0] | ((msg)->data[1] << 8) | ((msg)->data[2] << 16) | ((msg)->data[3] << 24))
 #define GET_BYTES_48(msg) ((msg)->data[4] | ((msg)->data[5] << 8) | ((msg)->data[6] << 16) | ((msg)->data[7] << 24))
 #define GET_FLAG(value, mask) (((__typeof__(mask))(value) & (mask)) == (mask))
